+++ tests/testthat/test-class_url.R
@@ -13,8 +13,14 @@ tar_test("dynamic urls work", {
     )
   })
   tar_make(callr_function = NULL)
-  exp <- tibble::tibble(name = "abc", progress = "built")
-  expect_equal(tar_progress(), exp)
+  exp <- tibble::tibble(
+    name = "abc",
+    type = "stem",
+    parent = "abc",
+    branches = 0L,
+    progress = "built"
+  )
+  expect_equal(tar_progress(fields = NULL), exp)
   tar_make(callr_function = NULL)
   expect_equal(nrow(tar_progress()), 0)
   meta <- tar_meta(abc)